refactor(cli): Extract common `prepareEnvironment` function from commands.
The new `prepareEnvironment` function discovers the Telestion Configuration file, extracts the config, and returns a Promise that resolves to the parsed configuration and path to the PSC's root directory.
