+++ packages/telestion-client-cli/src/commands/build.js
@@ -1,9 +1,8 @@
-const path = require('path');
-const getConfig = require('../lib/config');
 const compileElectronMainThread = require('../lib/build/compile-electron-main-thread');
 const compileReactApp = require('../lib/build/compile-react-app');
 const withAdjustedPackageJsonRun = require('../lib/build/with-adjusted-package-json-run');
 const packageElectronApp = require('../lib/build/package-electron-app');
+const prepareEnvironment = require('../lib/prepare-environment');
 const logger = require('../lib/logger')('build');
 
 // yargs def
@@ -18,16 +17,9 @@ function builder(yargs) {
 	});
 }
 
-async function handler(argv) {
+async function handler() {
 	try {
-		logger.info('Reading configuration');
-		const config = await getConfig();
-		logger.success('Found configuration file at: ' + config['filepath']);
-		logger.debug('Config', config);
-
-		let projectRoot = path.dirname(config['filepath']);
-		process.chdir(projectRoot);
-		logger.debug('Arguments:', argv);
+		let { config, projectRoot } = await prepareEnvironment();
 
 		logger.info('Compiling PSC React app');
 		await compileReactApp();
@@ -36,7 +28,7 @@ async function handler(argv) {
 		logger.info('Compiling PSC at', projectRoot);
 
 		logger.info('Compiling Electron main thread');
-		await compileElectronMainThread(projectRoot, config.config?.plugins || []);
+		await compileElectronMainThread(projectRoot, config?.plugins || []);
 		logger.success('Electron main thread has been compiled successfully.');
 
 		logger.info('Packaging Electron app');
